still debugging, trying to rework the algorithm

+++ src/distance_calculation.c
@@ -6,7 +6,7 @@
 /*   By: pchateau <pchateau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/04 10:52:05 by pchateau          #+#    #+#             */
-/*   Updated: 2025/04/12 10:06:52 by pchateau         ###   ########.fr       */
+/*   Updated: 2025/04/12 13:09:31 by pchateau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
  */
 float	calculate_distance_between_player_and_plane(void)
 {
-	return ((SCREEN_WIDTH / 2) / tan(convert_degree_to_radian(FOV/2)));
+	return ((SCREEN_WIDTH / 2) / tanf(convert_degree_to_radian(FOV/2)));
 }
 
 /**
@@ -39,5 +39,5 @@ float	fix_fish_eye_effect(t_ray ray, t_player player)
 	float	angle;
 	
 	angle = convert_degree_to_radian(player.direction_angle - ray.direction_angle);
-	return (ray.smallest_distance * fabs(cos(angle)));
+	return (ray.smallest_distance * fabs(cosf(angle)));
 }
